Move file to requestIdSchema
